requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.0.1
torchvision==0.15.2
pytorch-lightning==2.0.6
opencv-python==4.8.0.74
Pillow==10.0.0
numpy==1.24.3
segmentation-models-pytorch==0.3.3